some big step with database

+++ app/src/main/java/com/example/mindmaper/CentralNode.java
@@ -6,7 +6,7 @@ public class CentralNode extends Node {
     private ArrayList<ChildNode> rightChildren;
     private ArrayList<ChildNode> leftChildren;
 
-    public CentralNode(int nodeId, int styleId, String mainText, String attachedText) {
+    public CentralNode(long nodeId, long styleId, String mainText, String attachedText) {
         super(nodeId,styleId,mainText,attachedText);
         rightChildren = new ArrayList<>();
         leftChildren = new ArrayList<>();
